Fleshing out the introduction and adding another chapter
Changed the structure of the report so that there is now a 'State of the
Art' section, and a 'Background' section.
Added some discursive parts to the introduction, discussing the
motivations behind the research.
Added/adjusted some of the information in the honeypots state-of-the-art
section, and outlined what I will discuss in the state-of-the-art
chapter.
Added a basic outline for the sections that will be included in the
'Background' chapter.